+++ lib/database/common.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
 import 'dart:io';
+import 'dart:isolate';
 
 import 'package:drift/drift.dart';
+import 'package:drift/isolate.dart';
 import 'package:drift/native.dart';
 import 'package:flutter/foundation.dart';
 import 'package:intl/intl.dart';
@@ -35,4 +37,48 @@ LazyDatabase openDbConnection(
 
     return NativeDatabase(file);
   });
+}
+
+Future<DriftIsolate> createDriftIsolate(String dbFileName) async {
+  // this method is called from the main isolate. Since we can't use
+  // getApplicationDocumentsDirectory on a background isolate, we calculate
+  // the database path in the foreground isolate and then inform the
+  // background isolate about the path.
+  final dir = await getApplicationDocumentsDirectory();
+  final path = p.join(dir.path, dbFileName);
+  final receivePort = ReceivePort();
+
+  await Isolate.spawn(
+    _startBackground,
+    _IsolateStartRequest(receivePort.sendPort, path),
+  );
+
+  // _startBackground will send the DriftIsolate to this ReceivePort
+  return await receivePort.first as DriftIsolate;
+}
+
+void _startBackground(_IsolateStartRequest request) {
+  // this is the entry point from the background isolate! Let's create
+  // the database from the path we received
+  final executor = NativeDatabase(File(request.targetPath));
+  // we're using DriftIsolate.inCurrent here as this method already runs on a
+  // background isolate. If we used DriftIsolate.spawn, a third isolate would be
+  // started which is not what we want!
+  final driftIsolate = DriftIsolate.inCurrent(
+    () => DatabaseConnection(executor),
+  );
+  // inform the starting isolate about this, so that it can call .connect()
+  request.sendDriftIsolate.send(driftIsolate);
+}
+
+// used to bundle the SendPort and the target path, since isolate entry point
+// functions can only take one parameter.
+class _IsolateStartRequest {
+  _IsolateStartRequest(
+    this.sendDriftIsolate,
+    this.targetPath,
+  );
+
+  final SendPort sendDriftIsolate;
+  final String targetPath;
 }