+++ lib/database/database.dart
@@ -881,3 +881,15 @@ class JournalDb extends _$JournalDb {
     return linesAffected;
   }
 }
+
+// TODO: examine why failing with index already exists
+JournalDb getJournalDb() {
+  return JournalDb.connect(
+    DatabaseConnection.delayed(
+      Future.sync(() async {
+        final isolate = await createDriftIsolate(journalDbFileName);
+        return isolate.connect();
+      }),
+    ),
+  );
+}