+++ lib/database/sync_db.dart
@@ -97,3 +97,14 @@ class SyncDatabase extends _$SyncDatabase {
   @override
   int get schemaVersion => 1;
 }
+
+SyncDatabase getSyncDatabase() {
+  return SyncDatabase.connect(
+    DatabaseConnection.delayed(
+      Future.sync(() async {
+        final isolate = await createDriftIsolate(syncDbFileName);
+        return isolate.connect();
+      }),
+    ),
+  );
+}