+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Enable isolate support in JournalDb, SyncDb, LoggingDb
+- Run SyncDb and LoggingDb in separate isolate (thread)
 
 ## [0.8.169] - 2022-10-11
 ### Changed: